+++ assets/javascripts/kitten/components/form/text-input-with-unit/test.js
@@ -1,0 +1,69 @@
+import React from 'react'
+import renderer from 'react-test-renderer'
+import { TextInputWithUnit } from './index'
+
+describe('<TextInputWithUnit />', () => {
+  let component
+
+  describe('by default', () => {
+    beforeEach(() => {
+      component = renderer.create(<TextInputWithUnit />).toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('with valid prop', () => {
+    beforeEach(() => {
+      component = renderer.create(<TextInputWithUnit valid />).toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('with error prop', () => {
+    beforeEach(() => {
+      component = renderer.create(<TextInputWithUnit error />).toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('with disabled prop', () => {
+    beforeEach(() => {
+      component = renderer.create(<TextInputWithUnit disabled />).toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('with error prop', () => {
+    beforeEach(() => {
+      component = renderer.create(<TextInputWithUnit error />).toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('with unit prop', () => {
+    beforeEach(() => {
+      component = renderer
+        .create(<TextInputWithUnit>€</TextInputWithUnit>)
+        .toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+})